requirements.txt
six
click>=6.7
python-telegram-bot>=7.0.1